Simplifies the mutex and changes naming to match convention found in other languages.
This makes the mutex have two methods, `acquire` and `release`.  Calling `acquire` will either get you the mutex or it will throw and calling `release` will simply release the mutex.  This makes its usage much simpler in the common case as you only need to call `acquire`, you don't need to check the result or do a `get` first.

**NOTE**: There were a few places in the code where it was acquiring a mutex even if the caller was whitelisted.  I assumed that these were mistakes and replaced them with the pattern of "either you are whitelisted or you acquire, never both" but perhaps there is a legitimate reason that functions have the contract acquire a mutex even when whitelisted?  An example of this can be seen in the old code in `collectFees.se`

There may be value in a future improvement in creating a macro for `if (!whitelisted): MUTEX.acquire()` but I wanted to keep this PR as simple as possible.