Miner readability improvements.
Extracted out a function to find a file by name which allows us to get rid of a couple of loops (and therefore nesting).

Also made it so it raises an error if a contract was referenced but not actually included.

Finally removed a block that was in a `recompile == 0` check since this code is in an `if(recompile)` block meaning that check would always fail.