Migrates `completeSets.se` to `CompleteSets.sol`.
Along the way fixes an outstanding bug where share prices weren't correctly calculated for scalar markets and were only correct for binary/categorical markets by happy coincidence.  The fix was a functional change that I originally intended on addressing in the future, which is to make it so share purchases are 1:1 with denomination tokens.  If you buy a 5 complete sets it will cost 5 indivisible units of denomination token and you will receive 5 indivisible shares of each outcome.  This is now the case no matter what the range of the market is.

Fixes a bug with `fundedRepSnapshot` where it would get bleed-over from previous tests if it wasn't used in a test first.  Pytest fixtures are evaluated lazily, even if they are `session` so we need to first reset the base fixture when we want to create an extended fixture.