Deletes collection libraries, topics and remaining Serpent code.
With the decision to require an Augur node, Topics can be deleted.  This has a cascading effect of allowing us to delete some other stuff, including some collection libraries.  In the process, I deleted any collection libraries that were unused and migrated the single remaining used collection library over to Solidity (in the form of a struct with extensions).

Migrates the placeholder ReportingEscapeHatch to Solidity.

Removes the last of the solidity code (some macros and miscellaneous tests).