Fixes some signatures and some bugs.
There were a few type signatures that were not correct and fixing them lead to finding some bugs, which were fixed along the way.

The primary bug fixed here is estimating gas for contracts that take constructor parameters.  Previously the constructor parameters would not be provided during the estimate gas call, but they would be provided during the contract instantiation.  This would result in an incorrect gas estimate that would end up failing.