Fixes integration tests and renames LegacyReputationToken.
The rename is to bring the naming convention in line with the rest of our contracts.

The fix was to wait for legacy reputation token to be approved before calling migrate reputation.  This is necessary because when we estimate the gas cost of migration, it must be _after_ approval otherwise it underestimates because the transaction fails.