Now with client-side signing!
Adds an environment variable to integration tests for private key.  This is the key that will be used to sign all transactions submitted to the chain, rather then relying on the chain having an account unlocked.

This ended up being a pretty big refactor along the way, but isolated to integration test code.  The biggest change is the new `libraries/ContractInterfaces.ts`.  This is hand coded at the moment, but it is designed so that it can easily be auto-generated from the contract ABI output.  Authoring such a code generator should be pretty easy as its just a matter of mapping solc types around and looking at stateMutability.  The big win from this pattern over the previous pattern is that we get dev-time type checking and IDE auto-complete for all of our contracts (once code generator is authored) in TS.

Note: that contract method names ending with `_` are local calls.  Contract methods that are stateMutability of `view` or `pure` only have this overload, while contract methods that are `payable` or `nonpayable` have two copies of the method, one with `_` suffix and one without so the caller can decide whether they want to get the result or run the call on-chain.  I'm open to ideas to better ways to express this.  I originally went with a function parameter but that got kind of ugly fast.

I also abstracted out account management, connection management and configuration.  This was mainly to remove some code duplication that I was introducing as a side effect of the private key changes.

At this point we are just using the basics of ethjs, none of the more complex stuff like contract creation and call mapping.  I also removed the ABIParser class as I find the new statically typeded contract accessors to be far superior.

Finally, I tightened up a number of type definition files for some of the ethjs libraries we are using, as well as solc.  At the moment there is quite a bit of type duplication that we should take care of, but it is not really causing trouble at the moment so I have put it off once again.

Also fixes a couple solc warnings from recent PRs.