Makes the deploy script skip uploading contracts that don't match the file name.
This prevents us from attempting to upload helper library contracts that are included in a file with the primary contract in that file.  An alternitave would be to continue uploading them, but restructure how we do lookup keys.  I'm not actually sure what is best, but this solution was simple and fixes a bug where we were uploading a contract but then clobbering it in our internal data structures with the next (`FillOrder.sol`).