Makes controller registry store contract name mapped to address, commit_hash, and file_sha.
This change allows us to make scripts that are much smarter in how they upload contracts and also allows reviewers to easily go from on-chain contracts to source code with some basic means of validation along the way.

At the moment, the scripts just set the commit/file hashes but the next step (future PR) is to make the script look to see if a matching contract is already uploaded and if so only upload/update contracts that haven't yet been uploaded or have changed (different file sha) since they were uploaded.

I chose to change the controller registry mapping to point at a contract rather than having multiple mappings for clarity.  I also changed the `registerContract` (previously `setValue`) function to be `onlyOwnerCaller` rather than `onlyWhitelistedCallers` since that is a more appropriate constraint.