Makes ContractDeployer do incremental uploads.
If the currently uploaded contract (as registered with Controller) has the same bytecode sha as the bytecode to be uploaded then skip uploading.  Similarly, skip whitelisting/initialization if it is already done for the target contract.  Contains some pretty significant refactoring along the way, recommend ignoring the changeset for `ContractDeployer` and instead just reading the code directly.

I also cleaned up some of the definition files that were duplicated.