Fixup/cleanup in the constants file.
Standardizing on using `ether` when we mean ether and 10**18 when we mean REP.  Also removed some exponents where we really meant "number ending with some zeros".

We should probably also include the denominating asset in the name of these variables, like `_IN_ATTOREP` so it is immediately clear to the reader what the bond is _intended_ to be denominated in (helps avoid bugs where we use a bond wrong).

Adjusted the bond sizes so they are all 10x each other (meaning 10% ROI minimum for any bond placer after the first) and added a fixme note about figuring out the proper floor for designated reporter stake size (I wonder if it needs to be 137.5 REP?)